Added flexible generation to frequency

+++ firm_ce/components/components.py
@@ -10,7 +10,7 @@ class Generator:
         self.fuel = str(generator_dict['fuel']) # Fuel type ### CHANGE TO Fuel OBJECT
         self.max_build = int(generator_dict['max_build'])  # MW/year
         self.min_build = int(generator_dict['min_build'])  # MW/year
-        self.capacity = int(generator_dict['initial_capacity'])  # MW
+        self.capacity = float(generator_dict['initial_capacity'])  # MW
         self.unit_type = str(generator_dict['unit_type'])
         self.cost = UnitCost(int(generator_dict['capex']),
                               float(generator_dict['fom']),
@@ -36,7 +36,7 @@ class Storage:
         self.id = id
         self.name = str(storage_dict['name'])
         self.node = str(storage_dict['node'])
-        self.power_capacity = int(storage_dict['initial_power_capacity'])  # MW
+        self.power_capacity = float(storage_dict['initial_power_capacity'])  # MW
         self.energy_capacity = int(storage_dict['initial_energy_capacity'])  # MWh
         self.duration = int(storage_dict['duration']) if int(storage_dict['duration']) > 0 else 0
         self.charge_efficiency = float(storage_dict['charge_efficiency'])  # %
